refactor: rename GenCore to Huey
