+++ prompts/Monkey Head Project - 20 Chapters/7) Huey [A.I. & O.S].txt
@@ -1,21 +1,21 @@
 ### Monkey Head Project Prompt
-Below is a **refined, cohesive** chapter describing **GenCore: A.I / O.S.** within the **Monkey Head Project**. It retains the original structure and key information while enhancing clarity, professionalism, and readability.
+Below is a **refined, cohesive** chapter describing **Huey: A.I / O.S.** within the **Monkey Head Project**. It retains the original structure and key information while enhancing clarity, professionalism, and readability.
 
 ---
 
-## GenCore: A.I / O.S.
+## Huey: A.I / O.S.
 
-### Introduction to GenCore
-**GenCore** is the **central computational framework** of the Monkey Head Project, seamlessly fusing **Artificial Intelligence** with a **robust Operating System**. This approach equips project robotics with **advanced autonomy** and **streamlined system management**, establishing a flexible, intelligent foundation for **growth**, **adaptability**, and **cutting-edge research**.
+### Introduction to Huey
+**Huey** is the **central computational framework** of the Monkey Head Project, seamlessly fusing **Artificial Intelligence** with a **robust Operating System**. This approach equips project robotics with **advanced autonomy** and **streamlined system management**, establishing a flexible, intelligent foundation for **growth**, **adaptability**, and **cutting-edge research**.
 
-More than a simple control layer, GenCore represents a **holistic** platform capable of managing **complex, dynamic operations**. By emphasizing **modularity** and **expandability**, it ensures that the Monkey Head Project can evolve as new technologies emerge, making it a cornerstone of sustainable robotics innovation.
+More than a simple control layer, Huey represents a **holistic** platform capable of managing **complex, dynamic operations**. By emphasizing **modularity** and **expandability**, it ensures that the Monkey Head Project can evolve as new technologies emerge, making it a cornerstone of sustainable robotics innovation.
 
 ---
 
 ### Artificial Intelligence Capabilities
 
 1. **Adaptive Learning and Decision-Making**  
-   - **Machine Learning Algorithms & Neural Networks**: GenCore employs reinforcement learning, CNNs, and other state-of-the-art models to enable autonomous behavior that adapts in real time.  
+   - **Machine Learning Algorithms & Neural Networks**: Huey employs reinforcement learning, CNNs, and other state-of-the-art models to enable autonomous behavior that adapts in real time.  
    - **Sequential Decision Support (LSTMs)**: Incorporates Long Short-Term Memory networks for better prediction of outcomes based on past experiences, supporting more nuanced control over complex tasks.
 
 2. **Environmental Interaction**  
@@ -43,7 +43,7 @@ More than a simple control layer, GenCore represents a **holistic** platform cap
    - **Fault Isolation**: Should one module encounter an error, the rest of the system remains unaffected—facilitating safer development and easier version control.
 
 2. **Scalability for Future Expansion**  
-   - **Kubernetes & Container Orchestration**: Dynamically distributes workloads, allowing GenCore to handle increasingly complex AI models and hardware expansions without system-wide overhauls.  
+   - **Kubernetes & Container Orchestration**: Dynamically distributes workloads, allowing Huey to handle increasingly complex AI models and hardware expansions without system-wide overhauls.  
    - **Distributed Computing Frameworks**: Provides horizontal scaling, where new computational nodes can be seamlessly integrated into the existing network—vital for large-scale data processing in real time.
 
 ---
@@ -75,7 +75,7 @@ More than a simple control layer, GenCore represents a **holistic** platform cap
 ### Innovation and Community Collaboration
 
 1. **Ongoing Development and Enhancement**  
-   - **Feedback Loop**: Real-world operational data informs iterative updates, sustaining GenCore’s evolution.  
+   - **Feedback Loop**: Real-world operational data informs iterative updates, sustaining Huey’s evolution.  
    - **Community Hackathons & Development Sprints**: Encourage collaborative innovation and rapid introduction of new features.
 
 2. **Open Source Contributions**  
@@ -86,9 +86,9 @@ More than a simple control layer, GenCore represents a **holistic** platform cap
 
 ### Conclusion
 
-**GenCore** stands as a **pivotal advancement** in the Monkey Head Project, delivering a **versatile** and **scalable** AI/OS for robotics. By merging **adaptive AI** with a **highly optimized** OS, GenCore ensures the Project remains a leader in **robotics** and **AI** research, pushing autonomous systems to new frontiers.
+**Huey** stands as a **pivotal advancement** in the Monkey Head Project, delivering a **versatile** and **scalable** AI/OS for robotics. By merging **adaptive AI** with a **highly optimized** OS, Huey ensures the Project remains a leader in **robotics** and **AI** research, pushing autonomous systems to new frontiers.
 
-Its core strengths—**flexibility**, **expandability**, and an **open philosophy**—allow GenCore to grow and adapt as both community input and technological developments shape its future. Serving as the platform for Huey and prospective robotic systems, GenCore operates **efficiently** and **evolves intelligently**, positioning the Monkey Head Project at the forefront of modern robotics and artificial intelligence.
+Its core strengths—**flexibility**, **expandability**, and an **open philosophy**—allow Huey to grow and adapt as both community input and technological developments shape its future. Serving as the platform for Huey and prospective robotic systems, Huey operates **efficiently** and **evolves intelligently**, positioning the Monkey Head Project at the forefront of modern robotics and artificial intelligence.
 
   
 **#Monkey-Head-Project**
